moved attribute pair code to separate module
